refactor(identity): align magic-link session with shared identity model (#770)
